refactor: extract IssueExtractionResponseParser from IssueExtractionService (#519)
Completes the IssueExtractionService decomposition begun in #432 (request
builder). Moves the response-parsing + issue-normalization half — response-
envelope decoding, schema-repair/alias JSON mapping, and normalization of
issues, reproduction steps, and screenshot annotations — verbatim into a
new IssueExtractionResponseParser with a single entry point,
parseResult(from:session:).

IssueExtractionService is now a lean ~126-line transport/orchestration
actor: it does the request (via IssueExtractionRequestBuilder), the HTTP
call, status mapping, and timeout, then hands the 2xx body to the parser.

The parser/normalizer slices were extracted together because they are
coupled through the payload types (parse produces the payload;
makeIssueExtractionResult normalizes it). Behavior is byte-compatible:

- Identical AppError messages for invalid-response, non-2xx, empty,
  refusal, and unparseable content.
- The response-envelope decode error still propagates UNMAPPED (only
  IssueExtractionPayload.parse failures get the "unexpected format"
  message) — preserving the original do/catch boundary.

Existing IssueExtractionServiceTests (which feed response JSON through the
public API) pass unchanged as the characterization guard; a new focused
IssueExtractionResponseParserTests suite pins the parser's own API
(alias-key + markdown-fence repair, refusal/empty/unparseable messages).
Full unit suite green; codex review found no divergence.

Closes #519
